Make delete_accesslog support a BLACKHOLE master->normal slave
The DELETEs sent to the log master don't replicate to the slave if the
master is ENGINE=BLACKHOLE and the slave is (obviously) not.  This checks
for that situation by testing whether an id exists on the master, and
if not, sends the DELETEs directly to the slave.

tracks internal commit 4ec8f5ab75da8db56ce44a17d68947367a7c94e6