Add css specific to tag bars
The normal way to add a css file to some set of pages is to add
an entry in the db and let getCSS do the work.  tagbar.js, though
is required in exactly the same places as firehose.css, so instead
of adding an additional GET, we use the glob mechanism to add
tagbar.css right into firehose.css.

This is my initial implementation of the css, and it seems too
complicated, hopefully Wes will be able to slim it down.  In
the meanwhile, I have to have _something_, so here it is.

We really should glob more of our css.

tracks internal commit 2251f8a44a48e9e08c3292a21cf1b10ade7d6a5d