jQuery convention: names of vars that hold jQuery objects should start with '$'
For example

  var $tag_el = $('.tag');
  var tag = $tag_el.text();

I picked this up on the jQuery list.  It's becoming the standard convention and
seems like a pretty good idea to me.  Please adopt it in your new code, and
update your old code when convenient.  A natural next step of this idea would
be to rename free-standing functions that return jQuery objects to begin with
'$' as well.  Not sure I'm willing to go that far; on the other hand, I don't
see that we have any free-standing functions that return jQuery objects at the
moment.

tracks internal commit ad7f17710e0598a30690fc491f9207c8a04e514c