Rename: s/split_if_string/list_as_array/; add counter-part, list_as_string
All the tagui code deals with lists of tags and/or commands.  Most of the code
doesn't care whether that list is represented by an array of strings or by a
single string delimited by whitespace (or, where order is unimportant, as a hash
of the individual strings).  It usually has to end up, though, as an array for
iteration or manipulation.  This function gives me the flexibility to write
functions that can be anywhere in the processing stack: who ever gets the list
first turns it into an array.  Everybody below benefits.

The new name, list_as_array, and the counter-part function, list_as_string, make
the model more obvious (at least to me).

tracks internal commit 692669b527eabdbfed3931b8b681953ae6d505a7