Add global and jQuery-specific list splitting functions 'separate'
jQuery and Perl have a similar 'grep' function applicable to arrays.  It returns
a new array comprising exactly the elements for which the supplied predicate
evaluates to true. The annoying part is that sometimes you want to do one thing
to all the elements that match, and a different thing to all those that don't
--- particularly annoying if the thing you want to do is send the filtered lists
to different places.

For arbitrary JS arrays, that problem can now be solved with 'separate':

	var passed, failed;
	[ passed, failed ] = separate(list, fn);

passed and failed are now lists comprising excactly the elements from the
original list that passed or failed, respectively, the test in fn.  E.g.,

	var evens, odds;
	[evens, odds] = separate([1, 2, 3, 4, 5], function(v){
		return !(v % 2)
	});

	// evens == [2, 4]; odds == [1, 3, 5]

I've provided similar functionality for jQuery by extending it with a new
function of the same name.  The single argument can be either a function,
as above, or else a jQuery selector:

	[ $good, $bad ] = $(expr).separate(expr)

	[ $articles, $others ] = $(expr).separate('.article')

tracks internal commit 07c70524269bbc4c46b4933ef3446710b5e51bc3