Smarter nearest_parent() -- :first _may_ be fast, but each+early exit _must_ be
jQuery's $.each and $().each functions support early termination.  If your
callback returns false, the loop is done.  One hopes that's how ':first' is
implemented internally ... but for this case, it's worth ensuring because the
consequences are horrible.  Imagine:

	$this.nearest_parent(':has(span)')

If the search for ':has(span)' is performed at each successive parent, and only
_then_ a ':first' selected, you've just done an imense amount of work
(potentially exponential relative to the depth of the tree).  Since this is
exactly the kind of search for which I require nearest_parent(), it's worth
avoiding the bad case.

tracks internal commit b7c2efc0589eb7348d405ed06f82d72ed18f912e