Teach firehose to hold focus around updates; autocomplete ESC=clear, SPACE=submit
Caveat: saving/restoring focus only applies to the tag input field.  If the
autocompleter menu is up, it will still drop, and drop the selection to the end.
This badness is in the autocompleter itself, and will take more time (perhaps
than it is worth) to sort it out.

tracks internal commit 1e8cd0463e437c2a91118d80dd654dd40e152243