Package now generates stem names slightly differently
Before this commit, stem names were simply lower-cased with underscores between
parts, e.g.:

	TagUI.Responder		=>	tagui_responder
	ArticleInfo		=>	articleinfo

This commit changes that behavior to introduce an additional underscore at
transitions from a lowercase character (and a few others) to an uppercase character, e.g.,:

	TagUI.Responder		=>	tag_ui_responder
	ArticleInfo		=>	article_info
	$Foo_Bar		=>	$foo_bar

I did this because 'articleinfo' was unacceptable to me.  The consequence is I
had to some occurrences of 'tagui' in the code to 'tag_ui'; but for consistency
I fixed _all_ uses of tagui to be separated, in the case of classes and filenames,
with a '-', as in 'tag-ui.js'; in identifiers 'tag_ui'.

I know this _looks_ like a big change, but it really isn't.  It's the fix to the
Package function itself in slash.util.js, and then the renames.  That's all.

tracks internal commit 036893b01e851e7e7fe901f8dd12aaceee50d6f0