Negative admin tag commands affect opposite tags
Now all negative admin tag commands (_, #, ## etc., and as
appropriate ^) affect all opposite tags equally and in the
same way.  This works generically for both regular opposite
tags ("!foo") and specials.  For ^, the opposite tags are
reduced in clout the same way (up to that moment, not beyond).
The clout-reduction effects of # etc. are not propagated to
those who tagged the opposite way.

tracks internal commit 4682e2d9fa63869433744d176dd4d55073225230