Add users.seclev index
For years, getAdmins() has been doing a table scan on users to pull
the list of users with seclev that qualifies them to be admins.
And for years, we've gotten away with it because each process caches
that data forever, so there have been relatively few such requests.
Well, the users table finally got big enough that we have to pay
the piper and put a key on the column.  There's a tiny performance
penalty every time a new user is created but not wedging the database
with a flood of table scans is kinda worth it.

tracks internal commit ad64f4db577fb891e327fa5667cfb911e8a1db2e