serialize_multiple() salts non-unique fieldnames in the course of building a query string
This is a new JavaScript function in common.js to be used in place of, e.g., $form.serialize().
For the example, instead say this: serialize_multiple($form).  In the general case, the output
will be identical to $form.serialize().  However, in the case where the jQuery built-in would
return a string with non-unique fieldnames (as it does for multi-select lists, for instance),
our function tacks some salt onto the name.  Example:

	$form.serialize():
		single=Single&multiple=Multiple&multiple=Multiple3&check=check2&radio=radio1

	serialize_multiple( $form ):
		single=Single&multiple1=Multiple&multiple2=Multiple3&check=check2&radio=radio1

This is useful if you don't ever name fields with trailing significant digits.

Not yet called, but I added the call as a comment

tracks internal commit 13ea0570aab83690e7ee19a519e696ec5afd506c