Change order of sphinx indexes
The _dist index listed its agents in order from youngest to oldest.
Pudge found a place in the docs that notes that for killlist to
work correctly, the "proper" search order is oldest to youngest.
That point in the docs gives an example using Query() in the API,
but, suspecting that the same rule applies for agents listed in
an index definition, I'm changing ours around and hoping this
fixes an killlist problem we've identified.

tracks internal commit 68cd14657883120bc5b30a67ef61eb8c77c6680e