Change order of sphinx indexes
The _dist index listed its agents in order from youngest to oldest.
Pudge found a place in the docs that notes that for killlist to
work correctly, the "proper" search order is oldest to youngest.
That point in the docs gives an example using Query() in the API,
but, suspecting that the same rule applies for agents listed in
an index definition, I'm changing ours around and hoping this
fixes an killlist problem we've identified.

tracks internal commit 189de71fc8bb14c69e9de9f5c2cea11cd369515b