Save show/hide state of slashboxes
This code belongs with the button.  Moved similar recent code for #toggleleft
to the template for its button, as well.  That gets rid of the '.live' hack we
added to make sure it worked for Rob's demo.

However... still need work with/from Tim to get the slashboxes for hose loads
that didn't start with slashboxes shown.

tracks internal commit 906e437e28d5c657d7ccd49d4e62867f3001502d