firehose_reorder: split out other tasks, tighten up
Reorder items in the firehose.  Were it not for the i2 ad, we could simply
re-insert the items individually in the correct order.  Because of the i2 ad, we

	- must not: re-execute scripts of the i2 ad (which cannot be avoided if
		we move it within the DOM);

	- must not: allow the i2 to jump out of view innapropriately;

	- should not: insert items between the i2 ad and the following article
		(they may be related).

Previous iteration used '>#firehose-xyz', this turns out to be noticably slower
than either filtering children, or just selecting by id with _no_ context.  Go
figure.  Used to check if earlier moves obviated later ones, but we only expect
to move 1 run, if any...

While this algorithm was a joy to implement, it exists only because of a rare
(possibly dead) bug where a handful of new articles that belong at the bottom of
the firehose could accidentally be inserted at the top.  When that bug is known
to be fixed, we should remove this function and never speak of it again (/cry).

tracks internal commit b7433be72adb7943f8d5ae971a30acd9fa9eb528