Centralize the code and selectors for identifying firehose-items and/or chrome
firehose-chrome: day-breaks, floating ads, loading messages, etc.
firehose-items: articles, stories, etc., i.e., the opposite of chrome

Why:
	- we have vague plans of unifying classes later
	- the base-selector for firehose-items is complex (and beginning
	  to spread through the code-base) making it harder to
		- understand intent
		- spot errors
		- find commonality
		- optimize jQuery searches
		- make "unusual" selections
	- we have other article lists to which current code might apply
	  if we fixed the underlying selectors appropriately
	- duplication: more code than just the key-bindings need to know
	  the first, next, last, current, etc., firehose-item

This commit adds a new hunk of code, but doesn't change any existing code to
call it.  It will end up replacing quite a bit of code, I believe.

The new function is fhitems(), which returns a jQuery selection of
firehose-items at a level of abstraction just above those ugly details.  It uses
the fastest selection scheme it can.  The returned selection is in document
order except in the rare (and intentional) case noted.  The core form takes
three parameters: root, op, and filter.  I provide some "syntactic sugar" for
the most common cases, though; and the _most_ common case is: no arguments, or
else with a simple string by which the results will be filtered.  This is sugar
for setting the filter parameter.  The following calls all do exactly what you
would think (well ... they do exactly what _I_ would think, anyway):

	fhitems()			=> all firehose-items
	fhitems('.currfh')
	fhitems(':first')
	fhitems(':last')
	fhitems(':hidden')
	fhitems(':nth(5)')		=> 0-based

Perhaps less obviously:
	fhitems(':even')		=> every other firehose-item, 0-based
	fhitems('.article')		=> all expanded firehose-items
	fhitems('.briefarticle')	=> all collapsed
	fhitems(':has(span.votedup)')	=> all nodded (not yet purged)

Totally un-intuitively:
	fhitems(true)			=> all firehose-items _and_ -chrome
	fhitems(false)			=> all (and only) firehose-chrome

Or, for full control, call fhitems() with a filter function instead of a filter
string.  The "sugar-free" forms would be, e.g.,

	fhitems({filter:':first'});

The strings 'next' or 'prev' are sugar for setting the "op" parameter, so:

	fhitems('next') => fhitems({op:'next'})

...which returns the firehose-item immediately following the current one.

fhitems(element) returns (a jQuery selection of) the firehose-item that is or
contains that element.  It will do the same for every element of a jQuery
selection; and this is the one case where the result is ordered by the input
rather than the document.  fhitems(null|undefined) returns an empty jQuery
selection so it is safe to say, e.g.,

	fhitems(event.originalTarget)

...knowing you will get back an empty selection in the undefined case. This form
is implicitly setting the "root" parameter.

That's it for the sugar.  In the explicit form, you can combine at will, e.g.,

	fhitems({ op:'prev', filter:'*', root:fhitems('.article:first')})

...which returns the firehose-item _or_ -chrome immediately preceding the first
expanded article.

	fhitems({ op:'siblings', filter:'.article' })

...which returns all the expanded firehose-items _except_ current.

That's everything about directly selecting with fhitems(); but that's not the
end of the story.

"Saved" selectors:

Calling fhitems() directly, as in all the examples above, returns a jQuery
selection.  However, if you use the operator new, you get a re-usable object
that can fetch an updated jQuery selection any time you wish, e.g.,

	var next=new fhitems('next'), last=new fhitems(':last');
	function step(){
		firehose_set_cur(next.select());
	}

...step() advances the current firehose-item each time it is called; we can
reverse its meaning though by replacing or re-initing next:

	next.init('prev') or next=new fhitems('prev')

...after which step() moves in the opposite direction.

Finally, you may still want to use the actual selector strings we've been trying
to hide.  We need them in the .live() calls, for instance. Using the "new" form,
fhitems() will build exactly what you want:

	$(new fhitems().selector).live('focus-article', ...

The inner call produces (as of this writing):

	"#firehoselist>div[id^=firehose-]:not(.daybreak)"

...though when actually performing the selection, it uses something more like:

	$root.children('div[id^=firehose-]:not(.daybreak)')

Note: the code itself is only about half the length of this commit message.

tracks internal commit c96c1621820a83e09ac47a359411a8348680d3b3