Teach view to understand a "hint" (in how.hint), another element to consider in positioning
view() will calculate the goal position as though it _first_ scrolled how.hint into view,
and _then_ scrolled to the target element.  It still scrolls directly the goal.  The hint
is used only for the calculation.  The idea is to pass something like the h3 of the
next firehose-item (when moving forward), or of the previous item (if collapsed and moving
backward), or the body tag-widget (if expanded)

tracks internal commit f82ed300c735f85477735ec3014f80821525b208