Slash::Apache::Shtml
New module to handle emitting (some) .shtml files.  Its behavior
is to do a very crude check for the include-virtual server-parsed
lines indicating the page's header and footer, and replace them
with a real header() and footer() call which includes content (like
menus) customized for a logged-in user.  We trust this is not a
huge burden on the servers due to (a) Varnish caching most non-logged-in
.shtml requests (the logged-in ones it doesn't check cache for
because of the User cookie and doesn't write to cache both because
of the cookie and because header() emits no-cache headers for
logged-in users) and more importantly (b) most links to .shtml pages
are gone by now anyway.

tracks internal commit a2a3115e15ef67e732adf94f4c8fc40391186271