BugFix: color-picker menu must come in front of ad
Three things influence visible z-order:

	stacking context:
		For any element, the closest parent with a z-index is that element's
		stacking context.  Any element that specifies a z-index becomes a
		stacking context for its children.

	z-index:
		Within a stacking context (only) elements with lesser z-index render
		behind those with greater; but z-index never lets an element "break out"
		of its stacking context.

	DOM order:
		For elements in the same stacking context and with equal z-index, the
		earlier elements render behind later.

This is why elements should never need z-index:1000000.  It's not about running
faster than the bear.  It's about running faster than your friends.

The fix for the menu z-order bug is two-fold.  First, just before we .show(),
move the menu in the DOM to the end of the body.  That gives it priority for
both DOM order, and puts it at the top-level of stacking contexts.  Second,
assign it a z-index > any sibling context.  We could do this dynamically, but we
already know that both the ad and the modal_cover have z-index:1000000, so we
can just use that +1.  Simplified now that we're using global coordinates; and
in any case, fine-tuning the menu location relative to the trigger must be done
in CSS.

tracks internal commit 3c94ff61edc0b57eab81902598d3301e0ca42de8