Flip sprite scheme: _add_ fallback-styles rather than removing (fixes Safari)
The moment <div style="background: url(...image...)"> is added to the document,
the browser starts fetching that image.  Sprites are meant to reduce fetches by
combining images.  We must not let the browser fetch the non-sprite form of an
image if we have (or will have) it in a sprite---falling back on the
single-image rule otherwise.

The server can't know (at least beyond initial page-load) which sprites the
client actually has; else we could always supply the perfect 'style="..."' rule
at template-time.  Instead, client-side JavaScript must somehow "fix-up" the DOM
to use available sprites.  No JS: no sprites.  We _were_ exploiting this: hiding
such <div>s within <noscript> elements.  The browser couldn't "see" the hidden
fallback style-attribute.  Where we had a corresponding sprite, we would delete
the fallback rule.  Then, sprite or not, strip off the <noscript> wrapper.

When JavaScript is available, browsers don't _render_ the content of a
<noscript> element.  Our scheme---and as far as I can tell, the W3C
standards---require the browser to _preserve_ that content.  Safari doesn't.
For this problem, then, <noscript> can only help us in the actual case of no
JavaScript.  We can't "hide" the fallback style-attribute.

We need that fallback style, but it can't be delivered as a style attribute, nor
anywhere inside a <noscript> element; nor should it be delivered as an (invalid)
custom attribute.  In our new scheme, it's simply hidden text.

	BEFORE:
		<div class="topicfoo" style="background: url(...)">
			...
		</div>

	AFTER:
		<div class="maybe-sprite topicfoo">
			<span class="no-sprite" style="display:none">background: url(...)</span>
			...
		</div>

Where previously we sought to _remove_ fallback rules, we now search for those
cases that _require_ them: recreating the style attributes of yester-algorithm
from the text content of the inner rule element.  In the course of this work, we
clean as we go: removing the inner <span>s and outer classes once they have
served their purpose.

In the old scheme, script-disabled users got images; in the new scheme, they
won't.  Here is where <noscript> can still help.  We _could_, perhaps
conditionally, fall back to a <style> protected by <noscript> (yes, <style> _is_
valid in the <body>; and in HTML5, by the way, can optionally be "scoped" to its
parent).

	POSSIBLE:
		<div class="maybe-sprite topicfoo">
			<span class="no-sprite" style="display:none">background: url(...)</span>
			<noscript><style type="text/css">.topicfoo { background: url(...) }</style></noscript>
			...
		</div>

We haven't done this yet because duplicating the style feels ugly to me.

As part of this change, factored sprite_rules( rules ) to additionally accept a
single rule-string or an array of rule-strings to unconditionally add to the
the global <style id="sprite-rules">.  Its primary application, though, is still
to test entries in a map (class => rule), adding only those rules that are
actually needed.

tracks internal commit 4bf55b4d5ca211440e1c238cf95e9865e9a4d98e