Ad placement (new/old): align to "slugs" instead of doing math
Slugs are place-holders for the pin-positions of the slashbox-column ad. They
are mostly sized and positioned by CSS.  When the ad would pin to that position:
instead of doing math, I just move it to align with the slug. #slug-Crown and
part of the footer.

So that the slugs don't prevent ad-clicks, I adjusted some z-index values.
Setting it to 'auto' on #bd gave me a stacking context to control relative z
between the ad and the slashboxes.  1 on #firehose gave me room to put the slugs
underneath it.

A little care was needed with respect to slashbox-dragging and jbeard's slashbox
"adder".

Fixed two bugs: the ad should start out .Empty (hidden), not .No (unpinned); and
because fix_ad_position was bound to an event, it magically got a parameter
unrelated to would_recurse.  The new simplified logic means this calling pattern
is gone.  Also disabled some (broken on Firefox) code in firehose.html where I
was calling postMessage.  That code was not intended to be live, yet.

tracks internal commit c1257825d89024440d64a71725100227e5a59a93