jquery.autogrow.js plugin: now triggers events before and after growing
The events are triggered on the textarea itself.  Passes new_height in pixels.

'autogrow-will-grow'    triggered when animation starts
'autogrow-did-grow'     triggered when animation completes

You would bind to -will-grow if you wanted to animate the motion of the
upload control; and to -did-grow if you just wanted to fix it at the end.

tracks internal commit 845630ac39994c747965390539addd7d56b2252f