Fix display_prefs template for messages
dsiplay_prefs was missing the titlebar and generalbody div.  Looks like
it was set up for a modal dialog box. Since we are not using these,
needed the other above fixes.

Also reworded the text to eliminate  the text about options that are
not in use.

Also deleted a large section that was already commented out.