Update for increased username and pass
Also added admin interface to password prefs page.  Fixed so that upon
return of setting another users password, page goes back to edit with
the user still on.

This allows for the the login.pl to handle both user and admin password
prefs.

Made Newsletter by default a setting from the vars table