fixes for slashboxes and homepage reset
Remove code for blocks we will not be using (userlogin, random blocks).

Added a sanity check to when loading the users slashboxes.
This test to see if a box is actually currently ok to show.  This fixes
an issue where a users has a custom list of boxes, and an admin
deactivates one of the boxes.  The deactivated box will still be on the
users slashboxes list until the next time the user sets their homage
settings.  The san check will kill it even if listed on the users list.

Defaults are loaded from getPortalsCommon.
All blocks are loaded from getSectionBlocks.

These two are used to set up the defaults (when users slashboxes is set
to “”.  Uses the default_block field for this and the hidden field.

getSectionBlocks also returns always_on so we can use this to turn on
any block that is not in a user’s slashbox if this is set.

Removed users dynamic blocks as they are unnecessary and were not
showing up in the edit interface anyway.

Also fixed moderation not being set when user did a homepage settings
reset.