Closed open Anchor elements
Added a close tag for the anchors in the report table.  I found this when I found the misplaced partial anchor tag in 'headMessage;misc;default' in pull request 184.

Unlike 184, this wasn't creating a problem b/c browsers automatically closed the anchor when it saw the TD close.  This is just cleaner and won't cause issues if the HTML is changed.