Update for new viewpoll template.
Instead of using a naked pollbooth process template, set it so we can
use the viewpoll template that wraps the pollbooth  with the proper
generalbody div and puts a proper title in place.

Removed stupid fancybox template call as it was not needed and adding a
second title and basically putting a slashbox in a slashbox.

Have the save sub output viewpoll instead of the stupid save poll.  Set
the title to Poll Saved to show what happened.