Use word-wrap: break-word to break long words
Use css word-wrap: break-word to break long words.  Also set overflow:
wrap as this is a revision to the css that takes over the word-wrap
functionality.  This is set on the body tag to get everything.

Removed the breakHtml_ifwhitefix calls in Data.pm as we no longer need
this. Also removed the domain shortening when adding the domain at the
end of urls.

word-wrap does not work in a filedset properly, so set the edit
templates not to put the previews in a fieldset so that everything
should work properly.  This required the previw_comm template to be
slightly changes so that the bottom border could be put on the preview
for the comment.