Temporary fix for the tid bug
Stopping slashizing links in admin.pl for now so &tid=N isn't there to get messed up in the first place. The bug is in Data.pm in the sub _link_to_slashlink somewhere but I can't find it and don't have time to really dig in to it today.