Remove illegal <p>-within-<span> nesting
<span> elements are only permitted to contain gramattically-defined-as-inline
elements, which excludes <p>s. Styling the <p> as inline isn't enough to get
around that restriction. Fortunately, the <p>s were never actually doing
anything but contain their contents, and there was nothing between the <span>
and the <p>'s contents, so we can remove the entire <p></p> layer in the nesting.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>