Added HTML name directives
Added HTML name directives so helpful links can be directly to the relevant FAQ item. The style I used was to use the first letter of each word in the FAQ item as an aggregate label. This should be easily extensible, and can sustain exceptions (as long as you catch them, of course.)