Fix: Ensure 'fortune' command is available in Docker container
The 'fortune' command, which is used to display a random fortune at the bottom of the page, was not being updated after moving the site to the new server. This issue was reported by fliptop.

Added the 'fortune' package to the list of system build dependencies in the Dockerfile. This ensures that the 'fortune' command is available in the Docker container, which is necessary for the new_motd.pl task to function.

Changes:
- Updated Dockerfile to include 'fortune' in the list of packages installed by apt-get.

- kolie